refactor: cosmos-orchestration-account.test.ts
