+++ packages/orchestration/test/examples/stake-ica.contract.test.ts
@@ -230,32 +230,3 @@ test('makeAccountInvitationMaker', async t => {
   t.truthy(vstorageEntry, 'vstorage account entry created');
   t.is(bootstrap.marshaller.fromCapData(JSON.parse(vstorageEntry!)), '');
 });
-
-test('CosmosOrchestrationAccount - not yet implemented', async t => {
-  const { bootstrap } = await commonSetup(t);
-  const { publicFacet } = await startContract(bootstrap);
-  const account = await E(publicFacet).makeAccount();
-  const mockChainAddress: ChainAddress = {
-    value: 'cosmos1test',
-    chainId: 'cosmoshub-4',
-    encoding: 'bech32',
-  };
-  const mockAmountArg: AmountArg = { value: 10n, denom: 'uatom' };
-
-  await t.throwsAsync(E(account).getBalances(), {
-    message: 'not yet implemented',
-  });
-  await t.throwsAsync(E(account).send(mockChainAddress, mockAmountArg), {
-    message: 'not yet implemented',
-  });
-  // XXX consider, positioning amount + address args the same for .send and .transfer
-  await t.throwsAsync(E(account).transfer(mockAmountArg, mockChainAddress), {
-    message: 'not yet implemented',
-  });
-  await t.throwsAsync(E(account).transferSteps(mockAmountArg, null as any), {
-    message: 'not yet implemented',
-  });
-  await t.throwsAsync(E(account).withdrawRewards(), {
-    message: 'Not Implemented. Try using withdrawReward.',
-  });
-});
